refactor: split the two Unrolling passes, add basic helper functions
